Update Shadow submodule pointer. (#1543)

+++ integration-test/shadow/shadow_system_test.c
@@ -378,9 +378,9 @@ static void eventCallback( MQTTContext_t * pContext,
 {
     ShadowMessageType_t messageType = ShadowMessageTypeMaxNum;
     const char * pThingName = NULL;
-    uint16_t thingNameLength = 0U;
+    uint8_t thingNameLength = 0U;
     const char * pShadowName = NULL;
-    uint16_t shadowNameLength = 0U;
+    uint8_t shadowNameLength = 0U;
     uint16_t packetIdentifier;
 
     ( void ) pContext;
